separate dom from other files
