+++ src/index.js
@@ -1,6 +1,10 @@
 import "./assets/css/style.css";
 import 'bootstrap'
 import  './dom.js'
+import './display'
+
+
+
