+++ src/remove.js
@@ -1,0 +1,6 @@
+ export default function removeTask(ele){
+  if(ele.classList.contains('delete')){
+    ele.parentElement.parentElement.remove();
+  }
+}
+
